requirements.txt
asgiref==3.2.10
Django==3.1
djangorestframework==3.15.1
pytz==2024.1
sqlparse==0.5.0
validate-docbr==1.10.0
wrapt==1.12.1
